reorganizing the main folder
