+++ Redundant code/loginFunc.py
