+++ Redundant code/search.py
