+++ Redundant code/testSearch.py
