+++ data/user network.csv
@@ -1,0 +1,4 @@
+userid,user address,user password,user phone ,Name,Age,Email
+1548,515huhu86gj,494556,12122,,,
+51533,ygklj4,6461,17546,,,
+,,,,ali,22,ali@gmail
