--- project.dash.yaml
@@ -1,22 +1,0 @@
-global:
-    mode: dark
-datasets:
-    - uuid: hello
-      content: >
-        [
-          ["documentation", 50],
-          ["prototype", 40],
-          ["frontend", 10],
-          ["backend", 5],
-          ["architecture", 25]
-        ]
-pages:
-    - properties:
-        margin: 10px
-      components:
-        - settings:
-            type: BARCHART
-            general:
-                title: workflow
-            lookup:
-                uuid: hello
