--- user.csv
@@ -1,3 +1,0 @@
-userid,user address,user email,user password,user phone 
-1548,515huhu86gj,malak@gmail.com,494556,12122
-51533,ygklj4,sdfg@gmail.com,6461,17546
